Ensure hub_url is configured before use
When attempting to parse a unconfigured value and AttributeError is
raised. The reason is that the stand in class does not provide the
string functionality urlparse is attempting to use. To ensure teh
configuration is loaded even if the hub_url as not been configured use
an empty string as default.